Fix regression when configuring Jetty 9.0.x

+++ modules/cpr/src/main/java/org/atmosphere/container/Jetty9AsyncSupportWithWebSocket.java
@@ -81,7 +81,7 @@ public class Jetty9AsyncSupportWithWebSocket extends Servlet30CometSupport {
                 if (isJetty91) {
                     m = policy.getClass().getMethod("setMaxTextMessageSize", new Class[]{int.class});
                 } else {
-                    m = policy.getClass().getMethod("setMaxMessageSize", new Class[]{int.class});
+                    m = policy.getClass().getMethod("setMaxMessageSize", new Class[]{long.class});
                 }
                 m.invoke(policy, Integer.parseInt(max));
             }
@@ -93,7 +93,7 @@ public class Jetty9AsyncSupportWithWebSocket extends Servlet30CometSupport {
                 if (isJetty91) {
                     m = policy.getClass().getMethod("setMaxBinaryMessageSize", new Class[]{int.class});
                 } else {
-                    m = policy.getClass().getMethod("setMaxMessageSize", new Class[]{int.class});
+                    m = policy.getClass().getMethod("setMaxMessageSize", new Class[]{long.class});
                 }
                 m.invoke(policy, Integer.parseInt(max));
             }
